+++ sdk-lib/src/main/java/cash/z/ecc/android/sdk/internal/transaction/TransactionEncoderImpl.kt
@@ -19,12 +19,12 @@ import cash.z.ecc.android.sdk.model.Zatoshi
  * behaving like a stateless API so that callers can request [createTransaction] and receive a
  * result, even though there are intermediate database interactions.
  *
- * @property rustBackend the instance of RustBackendWelding to use for creating and validating.
+ * @property backend the instance of RustBackendWelding to use for creating and validating.
  * @property repository the repository that stores information about the transactions being created
  * such as the raw bytes and raw txId.
  */
-internal class WalletTransactionEncoder(
-    private val rustBackend: Backend,
+internal class TransactionEncoderImpl(
+    private val backend: Backend,
     private val saplingParamTool: SaplingParamTool,
     private val repository: DerivedDataRepository
 ) : TransactionEncoder {
@@ -75,7 +75,7 @@ internal class WalletTransactionEncoder(
      * @return true when the given address is a valid z-addr
      */
     override suspend fun isValidShieldedAddress(address: String): Boolean =
-        rustBackend.isValidShieldedAddr(address)
+        backend.isValidShieldedAddr(address)
 
     /**
      * Utility function to help with validation. This is not called during [createTransaction]
@@ -86,7 +86,7 @@ internal class WalletTransactionEncoder(
      * @return true when the given address is a valid t-addr
      */
     override suspend fun isValidTransparentAddress(address: String): Boolean =
-        rustBackend.isValidTransparentAddr(address)
+        backend.isValidTransparentAddr(address)
 
     /**
      * Utility function to help with validation. This is not called during [createTransaction]
@@ -97,14 +97,14 @@ internal class WalletTransactionEncoder(
      * @return true when the given address is a valid ZIP 316 Unified Address
      */
     override suspend fun isValidUnifiedAddress(address: String): Boolean =
-        rustBackend.isValidUnifiedAddr(address)
+        backend.isValidUnifiedAddr(address)
 
     override suspend fun getConsensusBranchId(): Long {
         val height = repository.lastScannedHeight()
-        if (height < rustBackend.network.saplingActivationHeight) {
+        if (height < backend.network.saplingActivationHeight) {
             throw TransactionEncoderException.IncompleteScanException(height)
         }
-        return rustBackend.getBranchIdForHeight(height)
+        return backend.getBranchIdForHeight(height)
     }
 
     /**
@@ -132,9 +132,9 @@ internal class WalletTransactionEncoder(
 
         @Suppress("TooGenericExceptionCaught")
         return try {
-            saplingParamTool.ensureParams(rustBackend.saplingParamDir)
+            saplingParamTool.ensureParams(backend.saplingParamDir)
             Twig.debug { "params exist! attempting to send..." }
-            rustBackend.createToAddress(
+            backend.createToAddress(
                 usk,
                 toAddress,
                 amount.value,
@@ -154,9 +154,9 @@ internal class WalletTransactionEncoder(
     ): Long {
         @Suppress("TooGenericExceptionCaught")
         return try {
-            saplingParamTool.ensureParams(rustBackend.saplingParamDir)
+            saplingParamTool.ensureParams(backend.saplingParamDir)
             Twig.debug { "params exist! attempting to shield..." }
-            rustBackend.shieldToAddress(
+            backend.shieldToAddress(
                 usk,
                 memo
             )